Added PCRE support.
* lisp/pdf-info.el (pdf-info-query--transform-response): Put match
  face properties on the matched text of the line.

(pdf-info-search-string): The returned line now contains face
properties on the matched part.

(pdf-info-search-regexp): This function now searches for PCRE, and as
such the calling convention has changed to support the various flags
of the underlying engine.
(pdf-info-regexp-compile-flags, pdf-info-regexp-match-flags):
List of regexp flags.

* server/epdfinfo.c: (cmd_search_regexp): Use complex numeric flags of
  the new engine instead of the old boolean ones.  Now returns both
  the matched text and it's corresponding line.
  (cmd_search_string): Return both as well.
  (cmd_regexp_flags): New command returning regexp flags and their values.

* lisp/pdf-util.el (pdf-util-highlight-regexp-in-string): New function
for propertizing a string.
(pdf-util-edges-union):  Added minor optimization.