Check for NixOS after macOS
On Darwin at least, if `nix` is installed then `autobuild` will fail as it tries
to pull NixOS dependencies. Changing the order of the OS check is sufficient to
fix this, though a more robust check for NixOS may be in order, since `nix` can
be installed on any Unix-like (even Windows subsytems).