Fix `tempnam' compilation warning
The following compilation warning was preventing epdfinfo to be linked
properly when I tried to install the pdf-tools modules from Melpa on
a Fedora 31.  As a consequence, the `pdf-tools-install' was failing
and I could not successfully enable this package.

   server/epdfinfo.c:354: warning: the use of `tempnam' is dangerous,
   better use `mkstemp'

The `tempnam' function is obsolete since POSIX.1-2008.

This patch re-implements the epdfinfo `mktempfile' function using the
`mkstemp' function instead of the `tempnam' function.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>