'set-window-hscroll does not update image-mode hscroll
image-mode keeps it's own copy of hscroll and during redisplay it will
grab that value and use it. If you don't update image-mode's hscroll,
but directly set the window's hscroll you will (after triggering a redisplay)
find your pdf scrolled all the way to the left margin. Quite unfortunate
if you happen to be zoomed in. This fixes that by using image-mode's
method for updating hscroll